light map progress
- viewer can load and apply light map from file for custom levels (scene.ApplyLightMap)
- added option to toggle textures
- various sanity checks, some extra commenting
- use uppercase to access some shared content
- added 3rd party color quantization code
- some minor pvs struct tests
- refactored VertexArrayControl to just get a pointer to scene instead of the array.

+++ ctr-tools/CTRFramework/Code/shared/Helpers.cs
@@ -122,6 +122,12 @@ namespace CTRFramework.Shared
         {
             string backupName = $"{fileName}.bkp";
 
+            if (!File.Exists(fileName))
+            {
+                Helpers.Panic("Helpers", PanicType.Error, "cannot backup a file that doesnt exist!");
+                return;
+            }
+
             if (!File.Exists(backupName))
                 File.Copy(fileName, backupName);
         }
